Added self destruction
BOOOM
