+++ src/main/java/frc/robot/Robot.java
@@ -22,7 +22,6 @@ public class Robot extends TimedRobot {
     LiveWindow.disableAllTelemetry();
     Context.robotController = new RobotController();
     robotStartTime = System.currentTimeMillis()/1000.0;
-    Context.robotController.compressor.start();
 
     // camera = edu.wpi.first.cameraserver.CameraServer.getInstance().startAutomaticCapture();
     // camera.setVideoMode(PixelFormat.kMJPEG, Context.cameraWidth, Context.cameraHeight, Context.cameraFPS);
@@ -56,23 +55,12 @@ public class Robot extends TimedRobot {
   @Override
 
   public void teleopPeriodic() {
-    Context.robotController.shooterController.setDesiredVelocity(1.5);
-    
     Context.robotController.loopAll();
     
     double driverThrottle = Context.robotController.driverJoystick.getThrottle();
     double driverYaw = Context.robotController.driverJoystick.getYaw();
     
     Context.robotController.drivetrain.arcadeDrive(driverYaw, driverThrottle);
-
-    
-    if((Context.robotController.opticalLocalization.LeftMovementX != 0) || (Context.robotController.opticalLocalization.LeftMovementY !=0))
-    {
-      System.out.println("X: " + Context.robotController.opticalLocalization.LeftMovementX + " Y: " + Context.robotController.opticalLocalization.LeftMovementY);
-    }
-    //System.out.println(String.format("X: 0x%08X, Y:  0x%08X",Context.robotController.opticalLocalization.LeftMovementX, Context.robotController.opticalLocalization.LeftMovementY));
- 
-    Context.setWOFTargetColor();
     
   }
 }
